Conflicts:
	src/me/piebridge/bible/Bible.java

+++ src/me/piebridge/bible/Bible.java
@@ -132,6 +132,7 @@ public class Bible
 
     public boolean checkVersions() {
         if (!setDatabasePath()) {
+            checkInternalVersions();
             return false;
         }
         File path = getExternalFilesDirWrapper();
@@ -149,12 +150,18 @@ public class Bible
         checkVersion(path);
         Collections.sort(versions);
         if (versions.size() == 0) {
-            setDemoVersions();
-            unpacked = true;
-            checkVersion(mContext.getFilesDir());
+            checkInternalVersions();
         }
         mtime.put(path.getAbsolutePath(), path.lastModified());
         return true;
+    }
+
+    private void checkInternalVersions() {
+        if (!unpacked) {
+            setDemoVersions();
+            unpacked = true;
+        }
+        checkVersion(mContext.getFilesDir());
     }
 
     public boolean isDemoVersion(String version) {
@@ -171,7 +178,7 @@ public class Bible
             return false;
         }
         File file = getFile(version);
-        if (file == null || !file.exists() || !file.isFile()) {
+        if (file == null || !file.isFile()) {
             if ("".equals(databaseVersion)) {
                 return setDefaultVersion();
             } else {
@@ -424,9 +431,6 @@ public class Bible
     }
 
     private void setDemoVersions() {
-        if (unpacked) {
-            return;
-        }
         int demoVersion = PreferenceManager.getDefaultSharedPreferences(mContext).getInt("demoVersion", 0);
         int versionCode = 0;
         try {
@@ -448,8 +452,13 @@ public class Bible
             return false;
         }
         String version = PreferenceManager.getDefaultSharedPreferences(mContext).getString("version", null);
-        if (version != null && getPosition(TYPE.VERSION, version) < 0) {
-            version = null;
+        if (version != null) {
+            // check actual file
+            File file = getFile(version);
+            if (file == null || !file.isFile()) {
+                checkVersions();
+                version = null;
+            }
         }
         if (version == null && getCount(TYPE.VERSION) > 0) {
             version = get(TYPE.VERSION, 0);
@@ -776,7 +785,11 @@ public class Bible
 
     public boolean deleteVersion(String version) {
         File file = getFile(version);
-        return file.delete();
+        if (file != null && file.isFile()) {
+            return file.delete();
+        } return {
+            return false;
+        }
     }
 
     public void checkBibleData(boolean block) {
